HeurBot v20 * Dumb dropoffs * Ship full = 900 halite * Dumb spawn logic (limit 100 remaining turns) * Smart collisions in 2P (only if my count ships neighbourhood > enemy's)

+++ src/basetypes.h
@@ -166,7 +166,7 @@ using PlayerDropoffs = Container<MAX_DROPOFFS, Dropoff *>;
 class Player
 {
   public:
-    size_t id, halite;
+    int id, halite;
     Point spawn;
 
     int action;
@@ -175,12 +175,12 @@ class Player
     PlayerDropoffs dropoffs;
 
     Player() : id(0), halite(0), spawn(0, 0), action(0), ships(), dropoffs() {}
-    Player(size_t id, int x, int y, Dropoff *shipyard) : id(id), halite(0), spawn(x, y), action(0)
+    Player(int id, int x, int y, Dropoff *shipyard) : id(id), halite(0), spawn(x, y), action(0)
     {
         dropoffs.put(shipyard);
     }
 
-    void update(size_t halite)
+    void update(int halite)
     {
         this->action = 0;
         this->halite = halite;
